+++ lib/src/ui/has_scrolling.dart
@@ -1,0 +1,11 @@
+//Copyright (C) 2012 Sergey Akopkokhyants. All Rights Reserved.
+//Author: akserg
+
+part of dart_web_toolkit_ui;
+
+/**
+ * Implemented by widgets that support scrolling.
+ */
+abstract class HasScrolling implements HasHorizontalScrolling, HasVerticalScrolling, HasScrollHandlers, IsWidget {
+  
+}