+++ lib/ui.dart
@@ -101,6 +101,8 @@ part 'src/ui/html_table.dart';
 part 'src/ui/flex_table.dart';
 part 'src/ui/grid.dart';
 part 'src/ui/simple_layout_panel.dart';
+part 'src/ui/scroll_panel.dart';
+part 'src/ui/has_scrolling.dart';
 
 part 'src/ui/inline_label.dart';
 part 'src/ui/inline_html.dart';
@@ -122,6 +124,7 @@ part 'src/ui/impl/history_impl.dart';
 part 'src/ui/impl/text_box_impl.dart';
 part 'src/ui/impl/clipped_image_impl.dart';
 part 'src/ui/impl/element_mapper_impl.dart';
+part 'src/ui/impl/scroll_impl.dart';
 
 /**
  * Supports core functionality that in some cases requires direct support from